validations and tests #1

+++ src/main/java/eu/sshopencloud/marketplace/services/tools/ToolService.java
@@ -1,9 +1,13 @@
 package eu.sshopencloud.marketplace.services.tools;
 
 import eu.sshopencloud.marketplace.dto.tools.ToolCore;
+import eu.sshopencloud.marketplace.model.auth.User;
 import eu.sshopencloud.marketplace.model.items.ItemCategory;
+import eu.sshopencloud.marketplace.model.items.ItemComment;
+import eu.sshopencloud.marketplace.model.items.ItemContributor;
 import eu.sshopencloud.marketplace.model.tools.Software;
 import eu.sshopencloud.marketplace.model.tools.Tool;
+import eu.sshopencloud.marketplace.repositories.auth.UserRepository;
 import eu.sshopencloud.marketplace.repositories.tools.ToolRepository;
 import eu.sshopencloud.marketplace.services.DataViolationException;
 import eu.sshopencloud.marketplace.services.items.ItemContributorService;
@@ -15,16 +19,25 @@ import eu.sshopencloud.marketplace.services.vocabularies.CategoryService;
 import eu.sshopencloud.marketplace.services.vocabularies.ConceptDisallowedException;
 import eu.sshopencloud.marketplace.services.vocabularies.PropertyService;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.lang3.StringUtils;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Sort;
+import org.springframework.security.authentication.AnonymousAuthenticationToken;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
+import javax.persistence.EntityNotFoundException;
+import java.time.ZonedDateTime;
+import java.util.ArrayList;
+import java.util.List;
 import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
+@Slf4j
 public class ToolService {
 
     private final ToolRepository toolRepository;
@@ -41,23 +54,26 @@ public class ToolService {
 
     private final ItemRelatedItemService itemRelatedItemService;
 
-    private  final SearchService searchService;
+    private final SearchService searchService;
+
+    private final UserRepository userRepository;
 
     public PaginatedTools getTools(int page, int perpage) {
         Page<Tool> tools = toolRepository.findAll(PageRequest.of(page - 1, perpage, new Sort(Sort.Direction.ASC, "label")));
         for (Tool tool: tools) {
-            tool.setRelatedItems(itemRelatedItemService.getItemRelatedItems(tool.getId()));
-            tool.setOlderVersions(itemService.getOlderVersionsOfItem(tool));
-            tool.setNewerVersions(itemService.getNewerVersionsOfItem(tool));
-            itemService.fillAllowedVocabulariesForPropertyTypes(tool);
+            tool = complete(tool);
         }
         return new PaginatedTools(tools, page, perpage);
     }
 
     public Tool getTool(Long id) {
         Tool tool = toolRepository.getOne(id);
-        toolRepository.flush();
-        tool.setRelatedItems(itemRelatedItemService.getItemRelatedItems(id));
+        tool = complete(tool);
+        return tool;
+    }
+
+    private Tool complete(Tool tool) {
+        tool.setRelatedItems(itemRelatedItemService.getItemRelatedItems(tool.getId()));
         tool.setOlderVersions(itemService.getOlderVersionsOfItem(tool));
         tool.setNewerVersions(itemService.getNewerVersionsOfItem(tool));
         itemService.fillAllowedVocabulariesForPropertyTypes(tool);
@@ -83,7 +99,7 @@ public class ToolService {
         if (newTool.getPrevVersionId() != null) {
             Optional<Tool> prevVersion = toolRepository.findById(newTool.getPrevVersionId());
             if (!prevVersion.isPresent()) {
-                throw new DataViolationException("prevVersionId", newTool.getDescription());
+                throw new DataViolationException("prevVersionId", newTool.getPrevVersionId());
             }
             result.setPrevVersion(prevVersion.get());
         }
@@ -104,7 +120,20 @@ public class ToolService {
     public Tool createTool(ToolCore newTool) throws DataViolationException, ConceptDisallowedException {
         String toolTypeCode = categoryService.getToolCategoryCode(newTool.getToolType());
         Tool tool = validate(newTool, toolTypeCode);
-        // TODO change previous versions by older and newer versions
+        itemService.switchVersionForCreate(tool);
+        ZonedDateTime now = ZonedDateTime.now();
+        tool.setLastInfoUpdate(now);
+
+        // TODO don't allow creating without authentication (in WebSecurityConfig)
+        Authentication authentication =  SecurityContextHolder.getContext().getAuthentication();
+        log.debug(authentication.toString());
+        if (! (authentication instanceof AnonymousAuthenticationToken)) {
+            User user = userRepository.findUserByUsername(authentication.getName());
+            List<User> informationContributors = new ArrayList<User>();
+            informationContributors.add(user);
+            tool.setInformationContributors(informationContributors);
+        }
+
         tool = toolRepository.save(tool);
         if (itemService.isNewestVersion(tool)) {
             if (tool.getPrevVersion() != null) {
@@ -112,18 +141,64 @@ public class ToolService {
             }
             searchService.indexItem(tool);
         }
+        tool = complete(tool);
         return tool;
     }
 
-    public Tool updateTool(Long id, Tool newTool) {
-        // TODO check ID
-        newTool.setId(id);
-        Tool tool = toolRepository.save(newTool);
-        // TODO index in SOLR
+    public Tool updateTool(Long id, ToolCore newTool) throws DataViolationException, ConceptDisallowedException, DisallowedToolTypeChangeException {
+        if (!toolRepository.existsById(id)) {
+            throw new EntityNotFoundException("Unable to find " + Tool.class.getName() + " with id " + id);
+        }
+        Tool currentTool = toolRepository.getOne(id);
+        // change between software and service is not allowed
+        String toolTypeCode = currentTool.getToolType().getCode();
+        if (newTool.getToolType() != null) {
+            if (!toolTypeCode.equals(newTool.getToolType().getCode())) {
+                throw new DisallowedToolTypeChangeException(toolTypeCode, newTool.getToolType().getCode());
+            }
+        }
+        Tool tool = validate(newTool, toolTypeCode);
+        tool.setId(currentTool.getId());
+        tool.setComments(currentTool.getComments());
+        if (tool.getId().equals(tool.getPrevVersion().getId())) {
+            throw new DataViolationException("prevVersionId", tool.getPrevVersion().getId());
+        }
+        itemService.switchVersionForUpdate(tool);
+        ZonedDateTime now = ZonedDateTime.now();
+        tool.setLastInfoUpdate(now);
+
+        // TODO don't allow creating without authentication (in WebSecurityConfig)
+        Authentication authentication =  SecurityContextHolder.getContext().getAuthentication();
+        log.debug(authentication.toString());
+        if (! (authentication instanceof AnonymousAuthenticationToken)) {
+            User user = userRepository.findUserByUsername(authentication.getName());
+            if (tool.getInformationContributors() != null) {
+                if (!tool.getInformationContributors().contains(user)) {
+                    tool.getInformationContributors().add(user);
+                }
+            } else {
+                List<User> informationContributors = new ArrayList<User>();
+                informationContributors.add(user);
+                tool.setInformationContributors(informationContributors);
+            }
+        }
+
+        tool = toolRepository.save(tool);
+        if (itemService.isNewestVersion(tool)) {
+            if (tool.getPrevVersion() != null) {
+                searchService.removeItem(tool.getPrevVersion());
+            }
+            searchService.indexItem(tool);
+        }
+        tool = complete(tool);
         return tool;
     }
 
     public void deleteTool(Long id) {
+        // TODO don't allow deleting without authentication (in WebSecurityConfig)
+        if (!toolRepository.existsById(id)) {
+            throw new EntityNotFoundException("Unable to find " + Tool.class.getName() + " with id " + id);
+        }
         Tool tool = toolRepository.getOne(id);
         itemRelatedItemService.deleteRelationsForItem(tool);
         itemService.switchVersionForDelete(tool);
